add setter-based config manipulation
